# Refactor rendering