--- a/addons/rmsmartshape/assets/Anchor.svg.import
+++ b/addons/rmsmartshape/assets/Anchor.svg.import
@@ -5,6 +5,7 @@ type="CompressedTexture2D"
 uid="uid://b2h8c634u1120"
 path="res://.godot/imported/Anchor.svg-1c23ca67e353b480fc5d8186bb0064b9.ctex"
 metadata={
+"has_editor_variant": true,
 "vram_texture": false
 }