--- a/addons/rmsmartshape/assets/CenterView.svg.import
+++ b/addons/rmsmartshape/assets/CenterView.svg.import
@@ -5,6 +5,7 @@ type="CompressedTexture2D"
 uid="uid://byx10wt6pl1vo"
 path="res://.godot/imported/CenterView.svg-f5e9c7178a59ab586ba9d8fac61799bc.ctex"
 metadata={
+"has_editor_variant": true,
 "vram_texture": false
 }