--- a/addons/rmsmartshape/assets/InterpLinear.svg.import
+++ b/addons/rmsmartshape/assets/InterpLinear.svg.import
@@ -5,6 +5,7 @@ type="CompressedTexture2D"
 uid="uid://umo7sgb6w70s"
 path="res://.godot/imported/InterpLinear.svg-c9d61ab817f2809471482a31f06cda10.ctex"
 metadata={
+"has_editor_variant": true,
 "vram_texture": false
 }