--- a/addons/rmsmartshape/assets/icon_collision_polygon_2d.svg.import
+++ b/addons/rmsmartshape/assets/icon_collision_polygon_2d.svg.import
@@ -5,6 +5,7 @@ type="CompressedTexture2D"
 uid="uid://c3n1j5ybtgjit"
 path="res://.godot/imported/icon_collision_polygon_2d.svg-d52c533e615c0baed724848cec175fe6.ctex"
 metadata={
+"has_editor_variant": true,
 "vram_texture": false
 }