--- a/addons/rmsmartshape/assets/icon_curve_create.svg.import
+++ b/addons/rmsmartshape/assets/icon_curve_create.svg.import
@@ -5,6 +5,7 @@ type="CompressedTexture2D"
 uid="uid://qf0hmmvirgbn"
 path="res://.godot/imported/icon_curve_create.svg-24c8317cfb8ced81d4c832ca7b3310d5.ctex"
 metadata={
+"has_editor_variant": true,
 "vram_texture": false
 }