--- a/addons/rmsmartshape/assets/icon_curve_delete.svg.import
+++ b/addons/rmsmartshape/assets/icon_curve_delete.svg.import
@@ -5,6 +5,7 @@ type="CompressedTexture2D"
 uid="uid://ce2cbowctirhc"
 path="res://.godot/imported/icon_curve_delete.svg-9121c11be9a22040b2d567743d44ead0.ctex"
 metadata={
+"has_editor_variant": true,
 "vram_texture": false
 }