--- a/addons/rmsmartshape/assets/icon_curve_edit.svg.import
+++ b/addons/rmsmartshape/assets/icon_curve_edit.svg.import
@@ -5,6 +5,7 @@ type="CompressedTexture2D"
 uid="uid://crjuwuq4vp15w"
 path="res://.godot/imported/icon_curve_edit.svg-bd7d5e1af7c72f4525dff1d79ad0af94.ctex"
 metadata={
+"has_editor_variant": true,
 "vram_texture": false
 }