--- a/addons/rmsmartshape/assets/icon_editor_freehand.svg.import
+++ b/addons/rmsmartshape/assets/icon_editor_freehand.svg.import
@@ -5,6 +5,7 @@ type="CompressedTexture2D"
 uid="uid://da64iqyjipw4f"
 path="res://.godot/imported/icon_editor_freehand.svg-4eb5cac550480a5ffb50b9c62fba3af1.ctex"
 metadata={
+"has_editor_variant": true,
 "vram_texture": false
 }