--- a/addons/rmsmartshape/assets/icon_editor_handle.svg.import
+++ b/addons/rmsmartshape/assets/icon_editor_handle.svg.import
@@ -5,6 +5,7 @@ type="CompressedTexture2D"
 uid="uid://ccnyogs0qbq8b"
 path="res://.godot/imported/icon_editor_handle.svg-ace1b9f89612a442855e8ad0888fdf3e.ctex"
 metadata={
+"has_editor_variant": true,
 "vram_texture": false
 }