--- a/addons/rmsmartshape/assets/icon_editor_handle_add.svg.import
+++ b/addons/rmsmartshape/assets/icon_editor_handle_add.svg.import
@@ -5,6 +5,7 @@ type="CompressedTexture2D"
 uid="uid://ci73r1ncmyrrw"
 path="res://.godot/imported/icon_editor_handle_add.svg-7258971e5edcfe4396b7b9c27c0ba1cc.ctex"
 metadata={
+"has_editor_variant": true,
 "vram_texture": false
 }