--- a/addons/rmsmartshape/assets/icon_editor_handle_bezier.svg.import
+++ b/addons/rmsmartshape/assets/icon_editor_handle_bezier.svg.import
@@ -5,6 +5,7 @@ type="CompressedTexture2D"
 uid="uid://cycko7tpd5xk5"
 path="res://.godot/imported/icon_editor_handle_bezier.svg-b0150f05772e8974a66bc735765e0aeb.ctex"
 metadata={
+"has_editor_variant": true,
 "vram_texture": false
 }