--- a/addons/rmsmartshape/assets/icon_editor_handle_control.svg.import
+++ b/addons/rmsmartshape/assets/icon_editor_handle_control.svg.import
@@ -5,6 +5,7 @@ type="CompressedTexture2D"
 uid="uid://dgcshah05p0bm"
 path="res://.godot/imported/icon_editor_handle_control.svg-8383942d01a7ee5e839992ed99e53aaf.ctex"
 metadata={
+"has_editor_variant": true,
 "vram_texture": false
 }