--- a/addons/rmsmartshape/assets/icon_editor_handle_selected.svg.import
+++ b/addons/rmsmartshape/assets/icon_editor_handle_selected.svg.import
@@ -5,6 +5,7 @@ type="CompressedTexture2D"
 uid="uid://dh3vdfbsl5o65"
 path="res://.godot/imported/icon_editor_handle_selected.svg-356ef806610ee2a60b5838a989f58598.ctex"
 metadata={
+"has_editor_variant": true,
 "vram_texture": false
 }