--- a/addons/rmsmartshape/assets/icon_editor_position.svg.import
+++ b/addons/rmsmartshape/assets/icon_editor_position.svg.import
@@ -5,6 +5,7 @@ type="CompressedTexture2D"
 uid="uid://c18veogqnx5ht"
 path="res://.godot/imported/icon_editor_position.svg-1f6e1977293add549ed436510f8834c1.ctex"
 metadata={
+"has_editor_variant": true,
 "vram_texture": false
 }