--- a/addons/rmsmartshape/assets/icon_editor_snap.svg.import
+++ b/addons/rmsmartshape/assets/icon_editor_snap.svg.import
@@ -5,6 +5,7 @@ type="CompressedTexture2D"
 uid="uid://bojrgd0r131xh"
 path="res://.godot/imported/icon_editor_snap.svg-f0374ecd8cb313d78a2acb46b093a7f1.ctex"
 metadata={
+"has_editor_variant": true,
 "vram_texture": false
 }