--- a/addons/rmsmartshape/shape_renderer.gd.uid
+++ b/addons/rmsmartshape/shape_renderer.gd.uid
@@ -1,0 +1,1 @@
+uid://dq85c8wcy410i