--- a/addons/rmsmartshape/shapes/shape_render.gd.uid
+++ b/addons/rmsmartshape/shapes/shape_render.gd.uid
@@ -1,1 +1,0 @@
-uid://dp60lb46toye7