… .github/workflows/unit_tests.yml
… README.md
… addons/rmsmartshape/documentation/FAQ.md
… addons/rmsmartshape/shapes/shape_render.gd
… examples/sharp_corner_tapering/sharp_corner_tapering.tscn
… tests/scenes/test_corners.tscn
… tests/scenes/test_mirrored.tscn
… tests/scenes/test_short_edge.tscn