The initial idea in #172 did not work out as planned.
Details to this and to all other changes are documented in respective commit messages.

- Undo the `Node2D` -> `MeshInstance2D` shape type conversion
- Use the `RenderingServer` directly for creating render nodes and rendering meshes
- Implement mesh baking by storing meshes in an exported property
- Make shapes clickable in editor by overlaying an invisible `ColorRect`
- Improve performance by reducing data juggling, code cleanup, etc.
- Remove multi-threaded edge generation because spawning threads is more expensive than the workload
- Render nodes now have texture repeat enabled by default


Here are some rough numbers measured using the complex shape (left) in `test_corners.tscn` example and averaged over multiple runs.

- The render time itself (allocating render nodes, setting render properties, rendering meshes) went down from `0.8 ms` to `0.09 ms`.
- The time for a full update (`force_update()`) was reduced from `4.2 ms`  to `3.3 ms` by the new rendering approach and code optimization.
- The `4 ms` of a full update can be reduced to `0.25 ms` on first load thanks to mesh baking.

|     | Render time | Full update time | Full update time (baked) |
|-----|-------------|------------------|--------------------------|
| New | 0.0866 ms   | 3.297 ms         | 0.254 ms                 |
| Old | 0.7676 ms   | 4.214 ms         | N/A                      |


Closes #172
Closes #76
Closes #61