@mphe 2025-07-08
Thanks for testing! Fixed the two problems.

> In my project, areas load in 30-180 ms tops, and used to be 1-2 seconds (on a beefy laptop)

That is a huge improvement. Didn't think it was still that slow before.