@limbonaut 2025-07-08
> Didn't think it was still that slow before.

Lots of shapes – corners, tapers, curves, even edge shaders 😅 
The speed bump is really noticeable.

> Thanks for testing! Fixed the two problems.

Let's merge it then.